requirements.txt
dataclasses==0.6
pytest==7.3.0
unittest2==1.1.0
pylint==2.16.2
PyYAML==6.0
GitPython==3.1.31
ruamel.yaml==0.17.21
pytz==2021.3
